Move the MX color & vector headers to src/shaders
We've already stripped these headers of any MX-specific bits. They seem
generally useful, so let's add them to src/shaders to anybody can use
them.  It's not "OSL standard library" (in stdosl.h and described in the
OSL Language Specification) but they are the next level of goodness --
distributed with the OSL library as a token of our beneficence toward
all mankind.